Implement generic content delivery for Tools
