+++ src/components/Tools/Tools/Structure/CollectionOverview/EquippedContainer/EquippedCompanions/EquippedCompanions.tsx
