+++ src/components/Tools/Tools/Structure/CollectionOverview/StringTextField/StringTextField.tsx
