+++ src/views/Tools/Tools.tsx
@@ -1,16 +1,19 @@
-import { SubNavigationBar } from 'components/SubNavigationBar/SubNavigationBar';
-import { navigationDataTools } from './subMenuData';
-
+import { ContentRouter } from "components/CotentRouter/ContentRouter";
+import { toolsList } from "components/Tools/ToolsCollection/ToolsList";
 export const Tools = () => {
   return (
     <>
-      <SubNavigationBar navigationData={navigationDataTools} />
       <div className='container-dark'>
-        <div className='justify-center container-light'>
-          <h1 className='text-3xl font-bold text-center'>Tools</h1>
-          <p className='text-xl text-center'>Tools and useful gadgets for your Kittens best interest!</p>
+        <div className='flex justify-center container-light'>
+          <div className='flex flex-col justify-center items-center'>
+            <h1 className='text-3xl font-bold'>Tools</h1>
+          </div>
+        </div>
+        <div className="container-light flex flex-col gap-2">
+          <ContentRouter contentList={toolsList} />
         </div>
       </div>
+
     </>
 
 	);
