--- src/views/Tools/subMenuData.ts
@@ -1,15 +1,0 @@
-import { INavigationRoute } from 'types/INavigationRoute';
-
-export const navigationDataTools: INavigationRoute = {
-	links: [
-		{
-			displayText: 'Overview',
-			url: '/tools',
-		},
-
-		{
-			displayText: 'Collection Display',
-			url: '/tools/collection-display',
-		},
-	],
-};
